darkmode: rename to ThemeToggle
